lab6

+++ lab6/runtime.c
@@ -4,9 +4,9 @@
 #include <string.h>
 
 
-int *initArray(int size, int init)
+long *initArray(int size, int init)
 {int i;
- int *a = (int *)malloc(size*sizeof(int));
+ long *a = (long *)malloc(size*sizeof(long));
  for(i=0;i<size;i++) a[i]=init;
  return a;
 }
